Split view controller to separate file
